pyproject.toml
[project]
name = "renaming-products"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fast-bitrix24>=1.8.4",
    "fastapi>=0.115.11",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]
